Replace sys.exit() with exception
It's generally bad form to exit a program from deep within the logic. I think it
shows that this code has been ported from Perl where this was probably a die(),
which is rather common in perl libs (and equally horrible as you can't save your
program from dying if you use it as a library). Since we want this code to be
usable as a library we avoid this and instead throw an exception which can be
caught and recovered from!